requirements.txt
fastapi==0.136.1
httpx==0.28.1
litellm==1.83.14
pandas==3.0.2
pydantic-settings==2.14.0
pydantic==2.12.5
pyjwt[crypto]==2.12.1
python-dotenv==1.2.2
resend>=2.0.0
supabase==2.29.0
svix>=1.40.0
uvicorn[standard]==0.46.0
